+++ .travis.yml
@@ -13,9 +13,9 @@ matrix:
         - python: 3.6
           env: KERAS_BACKEND=tensorflow
         - python: 2.7
-          env: KERAS_BACKEND=tensorflow USE_TF_KERAS=1 PYTEST_IGNORE='--ignore=tests/test_doc_auto_generation.py --ignore=tests/keras_contrib/backend --ignore=tests/keras_contrib/utils/save_load_utils_test.py'
+          env: KERAS_BACKEND=tensorflow PIP_OPTION=--use-tf-keras PYTEST_IGNORE='--ignore=tests/test_doc_auto_generation.py --ignore=tests/keras_contrib/backend --ignore=tests/keras_contrib/utils/save_load_utils_test.py'
         - python: 3.6
-          env: KERAS_BACKEND=tensorflow USE_TF_KERAS=1 PYTEST_IGNORE='--ignore=tests/test_doc_auto_generation.py --ignore=tests/keras_contrib/backend --ignore=tests/keras_contrib/utils/save_load_utils_test.py'
+          env: KERAS_BACKEND=tensorflow PIP_OPTION=--use-tf-keras PYTEST_IGNORE='--ignore=tests/test_doc_auto_generation.py --ignore=tests/keras_contrib/backend --ignore=tests/keras_contrib/utils/save_load_utils_test.py'
         - python: 2.7
           env: KERAS_BACKEND=theano THEANO_FLAGS=optimizer=fast_compile
         - python: 3.6
@@ -44,7 +44,9 @@ install:
   - travis_retry conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
   - source activate test-environment
   - travis_retry pip install --only-binary=numpy,scipy,pandas numpy nose scipy h5py theano pytest pytest-pep8 pandas --progress-bar off
-  - pip install git+https://github.com/keras-team/keras.git --progress-bar off
+  - if [[ "$PIP_OPTION" == "" ]]; then
+      pip install git+https://github.com/keras-team/keras.git --progress-bar off;
+    fi
 
   # set library path
   - export LD_LIBRARY_PATH=$HOME/miniconda/envs/test-environment/lib/:$LD_LIBRARY_PATH
@@ -53,7 +55,7 @@ install:
       travis_retry conda install -q mkl mkl-service;
     fi
 
-  - pip install -e .[tests] --progress-bar off
+  - pip install -e .[tests] --progress-bar off $PIP_OPTION
 
   # install TensorFlow (CPU version).
   - pip install tensorflow==1.9 --progress-bar off
@@ -81,11 +83,6 @@ script:
   # set up keras backend
   - sed -i -e 's/"backend":[[:space:]]*"[^"]*/"backend":\ "'$KERAS_BACKEND'/g' ~/.keras/keras.json;
   - echo -e "Running tests with the following config:\n$(cat ~/.keras/keras.json)"
-
-  - if [[ "$USE_TF_KERAS" == "1" ]]; then
-      pip uninstall -y keras && python convert_to_tf_keras.py;
-    fi
-
   - if [[ "$TEST_MODE" == "PEP8_DOC" ]]; then
       PYTHONPATH=$PWD:$PYTHONPATH py.test --pep8 -m pep8 -n0 && py.test tests/test_documentation.py;
     else
