+++ keras_contrib/tests/optimizers.py
@@ -5,7 +5,7 @@ from keras_contrib.utils import test_utils
 from keras import optimizers
 from keras.models import Sequential
 from keras.layers import Dense, Activation
-from keras.utils.np_utils import to_categorical
+from keras.utils import to_categorical
 
 
 def get_test_data():
